+++ pages/site_sidebar.php
@@ -106,6 +106,7 @@
           <ul class="treeview-menu">
             <li><a href="sdc_housing.php"><i class="fa fa-home"></i> Housing</a></li>
             <li><a href="sdc_hosting.php"><i class="fa fa-server"></i> Hosting</a></li>
+            <li><a href="sdc_iaas.php"><i class="fa fa-cloud"></i> IAAS</a></li>
           </ul>
         </li>
       </ul>